move regex logic to make-url-regex
